+++ Controller/__init__.py
@@ -1,4 +1,4 @@
-from .MController import MController
+from ..Model.MController import MController
 
 # Define what is available when importing `model`
 __all__ = ["MController"]
